+++ packages/terra-functional-testing/src/terra-cli/wdio/index.js
@@ -206,6 +206,11 @@ const cli = {
         describe: 'Whether or not to automatically update all reference screenshots with the latest screenshots.',
         default: false,
       },
+      useRemoteReferenceScreenshots: {
+        type: 'boolean',
+        describe: 'A flag to download reference screenshots from a remote site for screenshot comparisons instead of using the local reference screenshots.',
+        default: process.env.USE_REMOTE_REFERENCE_SCREENSHOTS === 'true',
+      },
       useSeleniumStandaloneService: {
         type: 'boolean',
         describe: 'A flag to use the selenium standalone service instead of the selenium docker service.',